+++ src/Engine.php
@@ -41,7 +41,7 @@ use function file_exists;
 use Infection\Configuration\Configuration;
 use Infection\Console\ConsoleOutput;
 use Infection\Event\ApplicationExecutionFinished;
-use Infection\Event\EventDispatcher\EventDispatcher;
+use Infection\Event\EventDispatcher;
 use Infection\Mutant\MetricsCalculator;
 use Infection\Mutation\MutationGenerator;
 use Infection\Performance\Limiter\MemoryLimiter;
