+++ src/Event/Subscriber/CiMutationGeneratingConsoleLoggerSubscriber.php
@@ -33,7 +33,7 @@
 
 declare(strict_types=1);
 
-namespace Infection\Event\Listener;
+namespace Infection\Event\Subscriber;
 
 use Infection\Event\MutationGeneratingStarted;
 use Symfony\Component\Console\Output\OutputInterface;
