+++ src/Event/Subscriber/EventSubscriber.php
@@ -33,32 +33,15 @@
 
 declare(strict_types=1);
 
-namespace Infection\Tests\Event\EventDispatcher;
+namespace Infection\Event\Subscriber;
 
-use Infection\Event\EventDispatcher\EventDispatcher;
-use Infection\Tests\Fixtures\Event\NullSubscriber;
-use Infection\Tests\Fixtures\Event\UnknownEventSubscriber;
-use Infection\Tests\Fixtures\Event\UserEventSubscriber;
-use Infection\Tests\Fixtures\Event\UserWasCreated;
-use PHPUnit\Framework\TestCase;
-
-final class EventDispatcherTest extends TestCase
+/**
+ * @internal
+ */
+interface EventSubscriber
 {
-    public function test_it_triggers_the_subscribers_registered_to_the_event_when_dispatcher_an_event(): void
-    {
-        $userSubscriber = new UserEventSubscriber();
-
-        $dispatcher = new EventDispatcher();
-        $dispatcher->addSubscriber($userSubscriber);
-        $dispatcher->addSubscriber(new NullSubscriber());
-        $dispatcher->addSubscriber(new UnknownEventSubscriber());
-
-        // Sanity check
-        $this->assertSame(0, $userSubscriber->count);
-
-        $dispatcher->dispatch(new UserWasCreated());
-        $dispatcher->dispatch(new UserWasCreated());
-
-        $this->assertSame(2, $userSubscriber->count);
-    }
+    /**
+     * @return array<string, callable>
+     */
+    public function getSubscribedEvents(): array;
 }
