+++ src/Event/Subscriber/MutationTestingResultsLoggerSubscriber.php
@@ -33,9 +33,8 @@
 
 declare(strict_types=1);
 
-namespace Infection\Event\Listener;
+namespace Infection\Event\Subscriber;
 
-use Infection\Event\EventDispatcher\EventSubscriber;
 use Infection\Event\MutationTestingFinished;
 use Infection\Logger\MutationTestingResultsLogger;
 use Webmozart\Assert\Assert;
