+++ src/Process/Builder/SubscriberBuilder.php
@@ -40,12 +40,12 @@ use Infection\Console\OutputFormatter\DotFormatter;
 use Infection\Console\OutputFormatter\OutputFormatter;
 use Infection\Console\OutputFormatter\ProgressFormatter;
 use Infection\Differ\DiffColorizer;
-use Infection\Event\EventDispatcher\EventDispatcher;
-use Infection\Event\EventDispatcher\EventSubscriber;
+use Infection\Event\EventDispatcher;
 use Infection\Event\Subscriber\CiInitialTestsConsoleLoggerSubscriber;
 use Infection\Event\Subscriber\CiMutantCreatingConsoleLoggerSubscriber;
 use Infection\Event\Subscriber\CiMutationGeneratingConsoleLoggerSubscriber;
 use Infection\Event\Subscriber\CleanUpAfterMutationTestingFinishedSubscriber;
+use Infection\Event\Subscriber\EventSubscriber;
 use Infection\Event\Subscriber\InitialTestsConsoleLoggerSubscriber;
 use Infection\Event\Subscriber\MutantCreatingConsoleLoggerSubscriber;
 use Infection\Event\Subscriber\MutationGeneratingConsoleLoggerSubscriber;
