+++ tests/phpunit/Event/Subscriber/CiMutantCreatingConsoleLoggerSubscriberTest.php
@@ -33,11 +33,11 @@
 
 declare(strict_types=1);
 
-namespace Infection\Tests\Event\Listener;
+namespace Infection\Tests\Event\Subscriber;
 
-use Infection\Event\EventDispatcher\EventDispatcher;
-use Infection\Event\Listener\CiMutantCreatingConsoleLoggerSubscriber;
+use Infection\Event\EventDispatcher;
 use Infection\Event\MutantsCreatingStarted;
+use Infection\Event\Subscriber\CiMutantCreatingConsoleLoggerSubscriber;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Output\OutputInterface;
