+++ tests/phpunit/Event/Subscriber/CiMutationGeneratingConsoleLoggerSubscriberTest.php
@@ -33,11 +33,11 @@
 
 declare(strict_types=1);
 
-namespace Infection\Tests\Event\Listener;
+namespace Infection\Tests\Event\Subscriber;
 
-use Infection\Event\EventDispatcher\EventDispatcher;
-use Infection\Event\Listener\CiMutationGeneratingConsoleLoggerSubscriber;
+use Infection\Event\EventDispatcher;
 use Infection\Event\MutationGeneratingStarted;
+use Infection\Event\Subscriber\CiMutationGeneratingConsoleLoggerSubscriber;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Output\OutputInterface;
