+++ tests/phpunit/Fixtures/Event/NullSubscriber.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Infection\Tests\Fixtures\Event;
 
-use Infection\Event\EventDispatcher\EventSubscriber;
+use Infection\Event\Subscriber\EventSubscriber;
 
 final class NullSubscriber implements EventSubscriber
 {
